% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfbp_handgrenade}
\alias{cfbp_handgrenade}
\title{Data from multi-increment sampling at Ft. Richardson}
\format{
A data frame with columns:
\describe{
\item{row}{The variable is integer. The row where the sample was collected}
\item{column}{The variable is integer. The column where the sample was collected}
\item{hmx}{The variable is numeric. The amount of HMX in mg/kg}
\item{rdx}{The variable is numeric. The amount of RDX in mg/kg}
}
}
\source{
\url{https://serdp-estcp.org/content/download/5167/73264/file/ER-0628-FR.pdf}
}
\usage{
cfbp_handgrenade
}
\description{
Validation of Sampling Protocol and the Promulgation of Method Modifications for the Characterization of Energetic Residues on Military Testing and Training Ranges
}
\examples{
\dontrun{
cfbp_handgrenade
}
}
\keyword{datasets}
