% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dta_fpsally}
\alias{dta_fpsally}
\title{Phase 2 Final Report: Incremental Sampling Visual Sample Plan Validation Study at Firing-Point Sally}
\format{
A data frame with columns:
\describe{
\item{row}{The variable is numeric. The row where the sample was collected}
\item{column}{The variable is numeric. The column where the sample was collected}
\item{dnt24}{The variable is numeric. The amount of 2,4 DNT in mg / kg}
\item{ng}{The variable is numeric. The amount of NG in mg / kg}
}
}
\source{
\url{data-raw/dta_fpsally.pdf}
}
\usage{
dta_fpsally
}
\description{
The objective of this project is to develop and demonstrate cost-effective statistical sampling techniques that can be used to confidently estimate the magnitude and extent of energetic compounds (i.e., explosives and propellants) within wide-area DUs when using ISMs.
}
\examples{
\dontrun{
dta_fpsally
}
}
\keyword{datasets}
