% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdx_alan}
\alias{rdx_alan}
\title{Data obtained from an Email from Alan Hewitt}
\format{
A data frame with columns:
\describe{
\item{row}{The variable is integer. The row where the sample was collected}
\item{column}{The variable is integer. The column where the sample was collected}
\item{rdx}{The variable is numeric. The amount of RDX in mg/kg}
}
}
\source{
\url{Email}
}
\usage{
rdx_alan
}
\description{
Discrete samples on a grid to compare to DU wide MIS
}
\examples{
\dontrun{
rdx_alan
}
}
\keyword{datasets}
