% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tce_chuck}
\alias{tce_chuck}
\title{MULTI INCREMENT TCE Vadose-Zone Investigation}
\format{
A data frame with columns:
\describe{
\item{row}{The variable is integer. The row where the sample was collected}
\item{column}{The variable is character. The column where the sample was collected}
\item{tce}{The variable is numeric. The amount of TCE in mg/kg}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/pdf/10.1002/rem.20196}
}
\usage{
tce_chuck
}
\description{
MULTI INCREMENT and discrete sampling strategies were used to estimate the average concentration and the three‐dimensional distribution of TCE in a 3,300‐m3 zone composed of two decision units.
}
\examples{
\dontrun{
tce_chuck
}
}
\keyword{datasets}
